web/requirements.txt
Pillow==9.1.0
Pillow==10.1.0
Requests==2.31.0
selenium==4.15.2
